Live2d viewer (#102)
And this is the server side code for the live2d viewer. Depends on MagiCircles/BanGDream-Static#1.

Currently this just adds a field to the Card model for a live2d zip, and two pages for the standalone and ajax viewers. ~~We still need to add the ability to upload zips through the API, and some touchups for the ajax stuff, but I'm putting this up for early review since I'd like to have it ready in time for leaks tomorrow.~~ This is done now.

The viewer works pretty well in Chrome, Firefox, Safari desktop, and iOS 11.2. Unfortunately I didn't have an android phone to try it with.

Some test files:
[tomoe.zip](https://github.com/MagiCircles/BanGDream/files/1987614/tomoe.zip)
[ako.zip](https://github.com/MagiCircles/BanGDream/files/1987616/ako.zip)

Closes #24

* add live2d viewer page

* add more controls and localize templates

* add screenshot buttons to the templates

* make the card page l2d link a button

* don't translate Live2D, it's a proper noun

* fix viewer not loading from an existing modal

* expose model field in API (write only)

* Live2D: share same code for standalone and ajax page + small changes ↓

- Default behavior for files in models got fixed in MagiCircles, so modify existing fields instead of re-adding
- Change ajax URL /ajax/live2d_ajax/{id}/ -> /ajax/live2d/{id}/
- Add helpers for live2durls inside the model (and use that when showing the buttons)
- Make live2d settings button with main color
- "Critical" (= performs an action) buttons with main color
- Cosmetics (mostly from my IDE)

* Live2d viewer: Translate controls