Made special_band more flexible
While the changes of us not knowing what the special band for a song is unlikely, with just two extra words in template we can allow ourselves that extra flexibility.

Also I once accidentally bypassed the clean thing anyways, so I feel it is just an easier safeguard to do this anyways.