Make Members Collection Show 5️⃣ Per Line
With the recent-ish update to MagiCircles allowing more obscure per_line choices, we can now show a band per line!

Made page_size 25 for now since it needs to be a factor of 5 & we have 25 girls currently, and made character name size h3 so that it'll all fit on one line!