Allow "Versionless" Official Art and filter them properly
Mainy for CN art, but applicable elsewhere too.

Similarly to comics, it changes the value field to a Yes/No on whether the asset is "versioned". Defaults to Yes so that most assets can be left alone for this.

If set to No, any version query other than null will remove the asset, and in ItemView the images for an asset won't specify a version.